fix:fix naming conflict

+++ src/Seeed_vl53l0x.h
@@ -56,7 +56,7 @@
 
 
 typedef int            s32;
-typedef unsigned int   u32;
+typedef long unsigned int   u32;
 typedef short          s16;
 typedef unsigned short u16;
 typedef char           s8;
